Rename mock data
